Refactor: 중복확인 input 컴포넌트 수정 및 import 경로 재설정
